CSE-515-P3-master/requirements.txt
cvxopt==1.2.7
cycler==0.11.0
fonttools==4.28.2
imageio==2.12.0
joblib==1.1.0
kiwisolver==1.3.2
Levenshtein==0.16.0
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
rapidfuzz==1.8.3
scikit-image==0.17.2
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
